PLFM-6265 refactored OAuthClientAuthFilter to be separate from BasicAuthenticationFilter
