+++ services/repository/src/main/java/org/sagebionetworks/auth/filter/DockerRegistryAuthFilter.java
@@ -42,7 +42,7 @@ public class DockerRegistryAuthFilter extends BasicAuthenticationFilter {
 			HttpServletRequest httpRequest, HttpServletResponse httpResponse, 
 			FilterChain filterChain, Optional<UserNameAndPassword> credentials) throws IOException, ServletException {
 		if (credentials.isPresent() && !validCredentials(credentials.get())) {
-			rejectRequest(httpResponse, getInvalidCredentialsMessage());
+			filterHelper.rejectRequest(reportBadCredentialsMetric(), httpResponse, getInvalidCredentialsMessage());
 			return;
 		}
